CAMEL-20587 - Camel-Qdrant: Add a datatype for transforming langchain embeddings in qdrant objects
