+++ components/camel-qdrant/pom.xml
@@ -49,6 +49,12 @@
         </dependency>
 
         <dependency>
+            <groupId>dev.langchain4j</groupId>
+            <artifactId>langchain4j-core</artifactId>
+            <version>${langchain4j.version}</version>
+        </dependency>
+
+        <dependency>
             <groupId>io.qdrant</groupId>
             <artifactId>client</artifactId>
             <version>${qdrant-client-version}</version>
